LayerCollection.njoin function usage
