+++ cassandra/src/main/scala/geotrellis/spark/io/cassandra/CassandraRDDReader.scala
@@ -7,15 +7,11 @@ import geotrellis.spark.io.avro.{AvroEncoder, AvroRecordCodec}
 import geotrellis.spark.io.index.{IndexRanges, MergeQueue}
 import geotrellis.spark.util.KryoWrapper
 
-import scalaz.concurrent.{Strategy, Task}
-import scalaz.std.vector._
-import scalaz.stream.{Process, nondeterminism}
 import com.datastax.driver.core.querybuilder.QueryBuilder
 import com.datastax.driver.core.querybuilder.QueryBuilder.{eq => eqs}
 import org.apache.avro.Schema
 import org.apache.spark.SparkContext
 import org.apache.spark.rdd.RDD
-import java.util.concurrent.Executors
 
 import com.typesafe.config.ConfigFactory
 
@@ -59,37 +55,25 @@ object CassandraRDDReader {
       .mapPartitions { partition: Iterator[Seq[(Long, Long)]] =>
         instance.withSession { session =>
           val statement = session.prepare(query)
-          val pool = Executors.newFixedThreadPool(threads)
 
           val result = partition map { seq =>
-            val range: Process[Task, Iterator[Long]] = Process.unfold(seq.toIterator) { iter =>
+            LayerReader.njoin[K, V](seq.toIterator, { iter =>
               if (iter.hasNext) {
-                val (start, end) = iter.next()
-                Some((start to end).toIterator, iter)
+                val index = iter.next()
+                val row = session.execute(statement.bind(index.asInstanceOf[java.lang.Long]))
+                if (row.nonEmpty) {
+                  val bytes = row.one().getBytes("value").array()
+                  val recs = AvroEncoder.fromBinary(kwWriterSchema.value.getOrElse(_recordCodec.schema), bytes)(_recordCodec)
+                  if (filterIndexOnly) Some(recs, iter)
+                  else Some(recs.filter { row => includeKey(row._1) }, iter)
+                } else Some(Vector.empty, iter)
               } else None
-            }
-
-            val read: Iterator[Long] => Process[Task, Vector[(K, V)]] = { iterator =>
-              Process.unfold(iterator) { iter =>
-                if (iter.hasNext) {
-                  val index = iter.next()
-                  val row = session.execute(statement.bind(index.asInstanceOf[java.lang.Long]))
-                  if (row.nonEmpty) {
-                    val bytes = row.one().getBytes("value").array()
-                    val recs = AvroEncoder.fromBinary(kwWriterSchema.value.getOrElse(_recordCodec.schema), bytes)(_recordCodec)
-                    if (filterIndexOnly) Some(recs, iter)
-                    else Some(recs.filter { row => includeKey(row._1) }, iter)
-                  } else Some(Vector.empty, iter)
-                } else None
-              }
-            }
-
-            nondeterminism.njoin(maxOpen = threads, maxQueued = threads) { range map read }(Strategy.Executor(pool)).runFoldMap(identity).unsafePerformSync
+            }, threads)
           }
 
           /** Close partition session */
           (result ++ Iterator({
-            pool.shutdown(); session.closeAsync(); session.getCluster.closeAsync(); Seq.empty[(K, V)]
+            session.closeAsync(); session.getCluster.closeAsync(); Seq.empty[(K, V)]
           })).flatten
         }
       }
