+++ s3/src/main/scala/geotrellis/spark/io/s3/S3RDDReader.scala
@@ -53,42 +53,25 @@ trait S3RDDReader {
     sc.parallelize(bins, bins.size)
       .mapPartitions { partition: Iterator[Seq[(Long, Long)]] =>
         val s3client = _getS3Client()
-        val pool = Executors.newFixedThreadPool(threads)
 
-        val result = partition map { seq =>
-          val range: Process[Task, Iterator[Long]] = Process.unfold(seq.toIterator) { iter =>
+        partition flatMap { seq =>
+          LayerReader.njoin[K, V](seq.toIterator, { iter =>
             if (iter.hasNext) {
-              val (start, end) = iter.next()
-              Some((start to end).toIterator, iter)
-            } else None
-          }
+              val index = iter.next()
+              val path = keyPath(index)
+              val getS3Bytes = () => IOUtils.toByteArray(s3client.getObject(bucket, path).getObjectContent)
 
-          val read: Iterator[Long] => Process[Task, Vector[(K, V)]] = { iterator =>
-            Process.unfold(iterator) { iter =>
-              if (iter.hasNext) {
-                val index = iter.next()
-                val path = keyPath(index)
-                val getS3Bytes = () => IOUtils.toByteArray(s3client.getObject(bucket, path).getObjectContent)
-
-                try {
-                  val bytes: Array[Byte] = getS3Bytes()
-                  val recs = AvroEncoder.fromBinary(kwWriterSchema.value.getOrElse(_recordCodec.schema), bytes)(_recordCodec)
-                  if (filterIndexOnly) Some(recs, iter)
-                  else Some(recs.filter { row => includeKey(row._1) }, iter)
-                } catch {
-                  case e: AmazonS3Exception if e.getStatusCode == 404 => Some(Vector.empty, iter)
-                }
-              } else {
-                None
+              try {
+                val bytes: Array[Byte] = getS3Bytes()
+                val recs = AvroEncoder.fromBinary(kwWriterSchema.value.getOrElse(_recordCodec.schema), bytes)(_recordCodec)
+                if (filterIndexOnly) Some(recs, iter)
+                else Some(recs.filter { row => includeKey(row._1) }, iter)
+              } catch {
+                case e: AmazonS3Exception if e.getStatusCode == 404 => Some(Vector.empty, iter)
               }
-            }
-          }
-
-          nondeterminism.njoin(maxOpen = threads, maxQueued = threads) { range map read }(Strategy.Executor(pool)).runFoldMap(identity).unsafePerformSync
+            } else None
+          }, threads)
         }
-
-        /** Close partition pool */
-        (result ++ Iterator({ pool.shutdown(); Vector.empty[(K, V)] })).flatten
       }
   }
 }
