+++ spark/src/main/scala/geotrellis/spark/io/file/FileRDDReader.scala
@@ -8,16 +8,12 @@ import geotrellis.spark.io.avro.{AvroEncoder, AvroRecordCodec}
 import geotrellis.spark.util.KryoWrapper
 import geotrellis.util.Filesystem
 
-import scalaz.concurrent.{Strategy, Task}
-import scalaz.std.vector._
-import scalaz.stream.{Process, nondeterminism}
 import org.apache.avro.Schema
 import org.apache.spark.SparkContext
 import org.apache.spark.rdd.RDD
 import com.typesafe.config.ConfigFactory
 
 import java.io.File
-import java.util.concurrent.Executors
 
 object FileRDDReader {
   def read[K: AvroRecordCodec: Boundable, V: AvroRecordCodec](
@@ -41,40 +37,24 @@ object FileRDDReader {
     val boundable = implicitly[Boundable[K]]
     val includeKey = (key: K) => KeyBounds.includeKey(queryKeyBounds, key)(boundable)
     val _recordCodec = KeyValueRecordCodec[K, V]
-    val kwWriterSchema = KryoWrapper(writerSchema) //Avro Schema is not Serializable
+    val kwWriterSchema = KryoWrapper(writerSchema) // Avro Schema is not Serializable
 
     sc.parallelize(bins, bins.size)
       .mapPartitions { partition: Iterator[Seq[(Long, Long)]] =>
-        val pool = Executors.newFixedThreadPool(threads)
-
-        val result: Iterator[Vector[(K, V)]] = partition map { seq =>
-          val range: Process[Task, Iterator[Long]] = Process.unfold(seq.toIterator) { iter =>
+        partition flatMap { seq =>
+          LayerReader.njoin[K, V](seq.toIterator, { iter =>
             if (iter.hasNext) {
-              val (start, end) = iter.next()
-              Some((start to end).toIterator, iter)
+              val index = iter.next()
+              val path = keyPath(index)
+              if (new File(path).exists) {
+                val bytes: Array[Byte] = Filesystem.slurp(path)
+                val recs = AvroEncoder.fromBinary(kwWriterSchema.value.getOrElse(_recordCodec.schema), bytes)(_recordCodec)
+                if (filterIndexOnly) Some(recs, iter)
+                else Some(recs.filter { row => includeKey(row._1) }, iter)
+              } else Some(Vector(), iter)
             } else None
-          }
-
-          val read: Iterator[Long] => Process[Task, Vector[(K, V)]] = { iterator =>
-            Process.unfold(iterator) { iter =>
-              if (iter.hasNext) {
-                val index = iter.next()
-                val path = keyPath(index)
-                if(new File(path).exists) {
-                  val bytes: Array[Byte] = Filesystem.slurp(path)
-                  val recs = AvroEncoder.fromBinary(kwWriterSchema.value.getOrElse(_recordCodec.schema), bytes)(_recordCodec)
-                  if (filterIndexOnly) Some(recs, iter)
-                  else Some(recs.filter { row => includeKey(row._1) }, iter)
-                } else Some(Vector.empty, iter)
-              } else None
-            }
-          }
-
-          nondeterminism.njoin(maxOpen = threads, maxQueued = threads) { range map read }(Strategy.Executor(pool)).runFoldMap(identity).unsafePerformSync
+          }, threads)
         }
-
-        /** Close partition pool */
-        (result ++ Iterator({ pool.shutdown(); Vector.empty[(K, V)] })).flatten
       }
   }
 }
